InputLogs.jsxがやや肥大していると感じたためより細かくComponentを分割
