+++ src/LearningLog.jsx
@@ -1,5 +1,6 @@
 import { useState } from 'react'
-import { InputLogs } from './components/InputLogs'
+import { FormGroups } from './components/FormGroups'
+import { PreviewList } from './components/PreviewList'
 import { LogList } from './components/LogList'
 
 
@@ -15,6 +16,11 @@ export const LearningLog = () => {
     setLogTime(e.target.value)
   }
 
+  const InputItems = [
+    { label: "学習項目", key: "title", type: "text", placeholder: "学習項目の概要を入力してください", value: logTitle, onChange: onChangeLogTitle },
+    { label: "学習時間", key: "time", type: "number", placeholder: "学習時間を入力してください", value: logTime, onChange: onChangeLogTime }
+  ]
+
   const records = [
     { title: "勉強の記録1", time: 1 },
     { title: "勉強の記録2", time: 3 },
@@ -23,12 +29,8 @@ export const LearningLog = () => {
   return (
     <>
       <h1>学習記録</h1>
-      <InputLogs
-        logTitle={logTitle}
-        logTime={logTime}
-        onChangeLogTitle={onChangeLogTitle}
-        onChangeLogTime={onChangeLogTime}
-      />
+      <FormGroups items={InputItems} />
+      <PreviewList items={InputItems} />
       <LogList logs={records} />
     </>
   )
